+++ Integrator/Model/ICarData.cs
@@ -6,6 +6,6 @@ using System.Text;
 namespace Integrator.Model {
     public interface ICarData {
         CarLocalization Localization { get; set; }
-        float Speed { get; set; }
+        double Speed { get; set; }
     }
 }
